Manual upload - setup git/eclipse better
The html has a web form that prompts for the user's name which is sent to the servlet on the web server after the user presses the submit button. Once inside the servlet it is used in the composition of the html formed in the servlet that is returned to the browser.